moved sorting to helpers
